refactor[#231] : features/map/hooks 폴더 생성
