+++ src/features/map/hooks/useResearchMarkingList.ts
